+++ src/main/java/ma/vi/esql/translation/MariaDbTranslator.java
@@ -71,12 +71,16 @@ public class MariaDbTranslator extends AbstractTranslator {
   }
 
   @Override
-  protected QueryTranslation translate(Update update, EsqlConnection esqlCon, EsqlPath path, PMap<String, Object> parameters, Environment env) {
+  protected QueryTranslation translate(Update               update,
+                                       EsqlConnection       esqlCon,
+                                       EsqlPath             path,
+                                       PMap<String, Object> parameters,
+                                       Environment          env) {
     StringBuilder st = new StringBuilder("update ");
 
     TableExpr from = update.tables();
     st.append(from.translate(target(), esqlCon, path.add(from), parameters, env));
-    Util.addSet(st, update.set(), target(), false, path);
+    Util.addSet(st, update.set(), target(), false, path, env);
 
     if (update.where() != null) {
       st.append(" where ").append(update.where().translate(target(), esqlCon, path.add(update.where()), parameters, env));